requirements.txt.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
pydantic-settings
aiorcon
rcon
